EventCode : int

+++ Assets/Scripts/GameManager.cs
@@ -91,7 +91,7 @@ public class GameManager : MonoBehaviourPunCallbacks, IOnEventCallback
     public void PropC() => RoleManager.propCount -= 1;
    public void OnEvent(EventData photonEvent)
 {
-    RoleManager.EventCodes eventCode = (RoleManager.EventCodes)photonEvent.Code;
+    RoleManager.EventCodes eventCode = (RoleManager.EventCodes)(int)photonEvent.Code;
     object[] data = (object[])photonEvent.CustomData;
 
     switch (eventCode)
